🔥 (cd/cd): no more latest branch
Updated the CICD action by removing the restrictions to the pushing
steps. Now when a PR is opened to "main" branch the images are built and
pushed (using branch name as tag) and when a commit is pushed to "main"
branch everiything is built and pushed (using "main" as tag).