👷 (ci/cd): update checkout method
Updated checkout method in order to be sure the action is working on
the head branch (the branch we want to merge into destination branch).
When a PR is closed could be because is merged or just closed.
If PR is merged, then a normal action/checkout will checkout the
destination branch (this will be main branch and we are going to untag
all the main images, not good).

With this commit we are forcing to checkout the head branch !