👷 (ci/cd): add action to untag images from closed PRs
This action will untag images with tags coming from closed PRs (merged
or closed). In this way we are forcing a specific branch (head branch,
aka source branch that we merged into destination branch) using '-e'
option to make command. In this way, regarding of the branch we are
checking out, we are removing always specific tag.